added snatched rasterisation algorithm

+++ src/config.rs
@@ -1,9 +1,9 @@
-pub const SCREEN_WIDTH: u16 = 320;
-pub const SCREEN_HEIGHT: u16 = 240;
+pub const SCREEN_WIDTH: usize = 320;
+pub const SCREEN_HEIGHT: usize = 240;
 
-pub const FB_TILE: u16 = 2;
+pub const FB_TILE: usize = 2;
 
-pub const FB_WIDTH: u16 = SCREEN_WIDTH / FB_TILE;
-pub const FB_HEIGHT: u16 = SCREEN_HEIGHT / FB_TILE;
+pub const FB_WIDTH: usize = SCREEN_WIDTH / FB_TILE;
+pub const FB_HEIGHT: usize = SCREEN_HEIGHT / FB_TILE;
 
 pub const TEST_N: usize = 50;
